requirements.txt

openai==1.28.1

Django==4.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
djoser==2.2.2
hashids==1.3.1
django-environ==0.11.2
dj-database-url==2.1.0
django-cloudinary-storage==0.3.0
python-decouple==3.8

# python3 -m venv venv  
# source venv/bin/activate
# install -r requirements.txt  